Исправила "комиссию", добавила custom validation

+++ src/main/java/faang/school/paymentservice/config/ExchangeCurrencyProperties.java
@@ -3,11 +3,13 @@ package faang.school.paymentservice.config;
 import faang.school.paymentservice.dto.Currency;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
+import java.math.BigDecimal;
+
 @ConfigurationProperties(prefix = "currency.exchange")
 public record ExchangeCurrencyProperties(
         String url,
         String appId,
         Currency base,
-        double commission
+        BigDecimal commission
 ) {
 }
